+++ apm-sniffer/optional-plugins/guava-cache-plugin/src/main/java/org/apache/skywalking/apm/plugin/guava/cache/GuavaCacheAllInterceptor.java
@@ -19,6 +19,7 @@
 package org.apache.skywalking.apm.plugin.guava.cache;
 
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
+import org.apache.skywalking.apm.agent.core.context.tag.Tags;
 import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
 import org.apache.skywalking.apm.agent.core.context.trace.SpanLayer;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
@@ -35,6 +36,7 @@ public class GuavaCacheAllInterceptor implements InstanceMethodsAroundIntercepto
         AbstractSpan span = ContextManager.createLocalSpan("GuavaCache/" + method.getName());
         span.setComponent(ComponentsDefine.GUAVA_CACHE);
         SpanLayer.asCache(span);
+        Tags.DB_TYPE.set(span, "GuavaCache");
     }
 
     @Override