+++ apm-sniffer/optional-plugins/guava-cache-plugin/src/main/java/org/apache/skywalking/apm/plugin/guava/cache/GuavaCacheInterceptor.java
@@ -39,6 +39,7 @@ public class GuavaCacheInterceptor implements InstanceMethodsAroundInterceptor {
         if (key != null) {
             Tags.DB_STATEMENT.set(span, key.toString());
         }
+        Tags.DB_TYPE.set(span, "GuavaCache");
         SpanLayer.asCache(span);
     }